Add video service layer, watch history, and UI components
Introduces VideoService and ImageService for handling video and image uploads, refactors video upload logic to use these services, and adds watch history tracking with a new WatchHistory model and migration. Adds Blade components for video cards and video grids, updates the home and channel pages to use these components, and improves video listing and playback logic. Updates migrations for better SQLite compatibility, adds support for manual video imports, and integrates text-to-speech functionality in the frontend. Also updates configuration and storage disk settings for new video workflows.

+++ database/migrations/2025_10_21_193138_create_likes_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('likes', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('video_id');
+            $table->unsignedBigInteger('user_id'); // SQLite uyumlu
+            $table->unsignedBigInteger('video_id'); // SQLite uyumlu
         });
     }
